Extracting Bulk Actions and WithSorting to Traits works.
